DRUP-1262: #ready fixes the data-label for location hours tables
This was due to javascript rendering the user's day on the date provided adjusted for their timezone. The desired day of the week is that actual day of that date. This was achieved by forcing javascript to use UTC for the timezone by passing true as the third parameter to the dateFormat() function. This is documented here: http://blog.stevenlevithan.com/archives/date-time-format